pyproject.toml
[project]
name = "nutritiondetector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
 "annotated-types==0.7.0",
 "anyio==4.8.0",
 "azure-functions==1.21.3",
 "certifi==2024.12.14",
 "charset-normalizer==3.4.1",
 "click==8.1.8",
 "fastapi==0.115.6",
 "h11==0.14.0",
 "httptools==0.6.4",
 "idna==3.10",
 "numpy==1.24.1",
 "peft>=0.14.0",
 "pillow==10.1.0",
 "pydantic-core==2.27.2",
 "pydantic-settings==2.7.1",
 "pydantic==2.10.5",
 "python-dotenv==1.0.1",
 "python-multipart>=0.0.20",
 "pyyaml==6.0.2",
 "requests==2.32.3",
 "sentencepiece==0.1.99",
 "sniffio==1.3.1",
 "starlette==0.41.3",
 "timm==0.9.10",
 "torch==2.1.2",
 "torchvision==0.16.2",
 "transformers==4.44.2",
 "typing-extensions==4.12.2",
 "urllib3==2.3.0",
 "uvicorn==0.34.0",
 "uvloop==0.21.0",
 "watchfiles==1.0.4",
 "websockets==14.2",
]
